fix:体力のデータ型をfloatに変更 refactor:Hitboxを衝突した相手を通知するクラスに変更

+++ Assets/Script/Combat/Health.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    [SerializeField] private int _maxHealth = 100;
-    [SerializeField] private int _currentHealth;
+    [SerializeField] private float _maxHealth = 100;
+    [SerializeField] private float _currentHealth;
     public float HealthRatio => _maxHealth == 0 ? 0 : Mathf.Clamp01((float)_currentHealth / _maxHealth);
 
     public event Action<float> OnHealthChanged;
@@ -27,7 +27,7 @@ public class Health : MonoBehaviour
     }
 #endif
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         if (amount <= 0) return;
